Try running integration tests during nix flake check.
This commit is more of a documentation, than the actual solution. I was
able to get to the point when things almost work, or at least the test
code worked, but we cannot run `nix flake lock`, because it would need
internet access which is not available during bulding a derivation.